% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{assign_zeros_where_absent_type}
\alias{assign_zeros_where_absent_type}
\title{assign zeros where habitat/geomorpho type is absent using habitat/geomorpho polygons}
\usage{
assign_zeros_where_absent_type(cor, mask)

assign_zeros_where_absent_type(cor, mask)
}
\arguments{
\item{cor}{}

\item{mask}{}

\item{r}{}

\item{rastxy}{}
}
\description{
assign zeros where habitat/geomorpho type is absent using habitat/geomorpho polygons

assign zeros where habitat/geomorpho type is absent using coral cover raster as mask
}
