% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve_prioritisation.R
\name{calc_reserve_patch_surfaces}
\alias{calc_reserve_patch_surfaces}
\title{Calculate reserve patch surfaces in km2}
\usage{
calc_reserve_patch_surfaces(reserves)
}
\arguments{
\item{reserve}{}
}
\description{
Calculate reserve patch surfaces in km2
}
