% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{calculate_median_slope}
\alias{calculate_median_slope}
\title{Calculate median slope from median depth}
\usage{
calculate_median_slope(dep)
}
\arguments{
\item{dep}{}
}
\description{
Calculate median slope from median depth
}
