% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve_prioritisation.R
\name{calibrate_blm_round1}
\alias{calibrate_blm_round1}
\title{Calibrate blm value for given target (minimum set objective, with penalty, lock in) round 1 (BLM from 0 to 100)}
\usage{
calibrate_blm_round1(pred, t)
}
\arguments{
\item{t}{}
}
\description{
Calibrate blm value for given target (minimum set objective, with penalty, lock in) round 1 (BLM from 0 to 100)
}
