% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve_prioritisation.R
\name{calibrate_blm_round2}
\alias{calibrate_blm_round2}
\title{Calibrate blm value for given target (minimum set objective, with penalty, lock in) round 2}
\usage{
calibrate_blm_round2(pred, t, blmval)
}
\arguments{
\item{t}{}
}
\description{
Calibrate blm value for given target (minimum set objective, with penalty, lock in) round 2
}
