% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_clean.R
\name{clean_off_effort_transit_wcoast}
\alias{clean_off_effort_transit_wcoast}
\title{Clean off effort portion related to transits west coast}
\usage{
clean_off_effort_transit_wcoast(off)
}
\arguments{
\item{off}{}
}
\description{
Clean off effort portion related to transits west coast
}
