% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve_prioritisation.R
\name{design_reserves_minset_no_penalty}
\alias{design_reserves_minset_no_penalty}
\title{Design reserve based on predicted densities with minimum set objective and given target with no penalty}
\usage{
design_reserves_minset_no_penalty(region, pred, target)
}
\arguments{
\item{region}{}

\item{pred}{}

\item{target}{}
}
\description{
Design reserve based on predicted densities with minimum set objective and given target with no penalty
}
