% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve_prioritisation.R
\name{design_reserves_minset_no_penalty_lockin}
\alias{design_reserves_minset_no_penalty_lockin}
\title{Design reserve based on predicted densities with minimum set objective and given target with locked in constraint on mpas}
\usage{
design_reserves_minset_no_penalty_lockin(region, pred, target, mpas)
}
\arguments{
\item{region}{}

\item{pred}{}

\item{target}{}

\item{mpas}{}
}
\description{
Design reserve based on predicted densities with minimum set objective and given target with locked in constraint on mpas
}
