% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve_prioritisation.R
\name{design_reserves_minset_no_penalty_randomize}
\alias{design_reserves_minset_no_penalty_randomize}
\title{Design reserve based on predicted densities with minimum set objective and given target with locked in constraint on mpas
with randomisation of uniform feature raster (corresponding to species range) - this returns a portofolio of solutions}
\usage{
design_reserves_minset_no_penalty_randomize(region, pred, target)
}
\arguments{
\item{region}{}

\item{pred}{}

\item{target}{}

\item{mpas}{}
}
\description{
Design reserve based on predicted densities with minimum set objective and given target with locked in constraint on mpas
with randomisation of uniform feature raster (corresponding to species range) - this returns a portofolio of solutions
}
