% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve_prioritisation.R
\name{design_reserves_minset_with_penalty}
\alias{design_reserves_minset_with_penalty}
\title{Design reserve based on predicted densities with minimum set objective and given target with penalty}
\usage{
design_reserves_minset_with_penalty(region, pred, target, penalty)
}
\arguments{
\item{region}{}

\item{pred}{}

\item{target}{}

\item{penalty}{}
}
\description{
Design reserve based on predicted densities with minimum set objective and given target with penalty
}
