% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{extend_raster}
\alias{extend_raster}
\title{Extend rasters based on defined study area raster}
\usage{
extend_raster(r, rast)
}
\arguments{
\item{r}{}

\item{rast}{}
}
\description{
Extend rasters based on defined study area raster
}
