% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{fit_full_abundance_model_old}
\alias{fit_full_abundance_model_old}
\title{fit full abudance GAM with area surveyed as offset and
unlimited k}
\usage{
fit_full_abundance_model_old(distrib, data_model)
}
\arguments{
\item{distrib}{}
}
\description{
fit full abudance GAM with area surveyed as offset and
unlimited k
}
