% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_density_rasters.R
\name{get_density_stack_species}
\alias{get_density_stack_species}
\title{Get raster stack of density for given species}
\usage{
get_density_stack_species(df, species)
}
\arguments{
\item{df}{}

\item{species}{}
}
\description{
Get raster stack of density for given species
}
