% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{get_polygon_surveyed_area}
\alias{get_polygon_surveyed_area}
\title{Get polygon of surveyed area}
\usage{
get_polygon_surveyed_area(r)
}
\arguments{
\item{r}{}
}
\description{
Get polygon of surveyed area
}
