% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{hist_predicted_abundance}
\alias{hist_predicted_abundance}
\title{Histogram of predicted abundances}
\usage{
hist_predicted_abundance(rast)
}
\arguments{
\item{name}{}
}
\description{
Histogram of predicted abundances
}
