% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_clean.R
\name{join_wcoast_video_info_telem}
\alias{join_wcoast_video_info_telem}
\title{Join west coast video information to telemetry}
\usage{
join_wcoast_video_info_telem(vid, telem)
}
\description{
Join west coast video information to telemetry
}
