% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{load_travel_time_raster}
\alias{load_travel_time_raster}
\title{Load original travel time raster (in seconds) from Florian Baletaud and resample}
\usage{
load_travel_time_raster(rast)
}
\arguments{
\item{rast}{}
}
\description{
Load original travel time raster (in seconds) from Florian Baletaud and resample
}
