% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_density_rasters.R
\name{make_abundance_df}
\alias{make_abundance_df}
\title{Make dataframe per grid cell centers (including empty cell centers) of abundance and surveyed area}
\usage{
make_abundance_df(polyobs, polytracks, footprintwidth)
}
\arguments{
\item{footprintwidth}{}
}
\description{
Make dataframe per grid cell centers (including empty cell centers) of abundance and surveyed area
}
