% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_clean.R
\name{make_area_raster_latlon}
\alias{make_area_raster_latlon}
\title{Make raster for study area in lat lon with given resolution}
\usage{
make_area_raster_latlon(lat1, lon1, lat2, lon2, res)
}
\arguments{
\item{res}{}
}
\description{
Make raster for study area in lat lon with given resolution
}
