% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_clean.R
\name{make_area_raster_xy}
\alias{make_area_raster_xy}
\title{Make study area raster projected to lambert New caledonia in meters with given resolution in meters}
\usage{
make_area_raster_xy(lat1, lon1, lat2, lon2, res_meters)
}
\arguments{
\item{res_meters}{}
}
\description{
Make study area raster projected to lambert New caledonia in meters with given resolution in meters
}
