% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{make_coral_cover_raster}
\alias{make_coral_cover_raster}
\title{Make coral cover raster}
\usage{
make_coral_cover_raster(coralpoly, ncraster)
}
\arguments{
\item{ncraster}{}
}
\description{
Make coral cover raster
}
