% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{make_dist_to_land_raster}
\alias{make_dist_to_land_raster}
\title{Make distance to land raster}
\usage{
make_dist_to_land_raster(coral, raster)
}
\arguments{
\item{raster}{}
}
\description{
Make distance to land raster
}
