% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{make_dist_to_pass_raster}
\alias{make_dist_to_pass_raster}
\title{Make distance to reef pass}
\usage{
make_dist_to_pass_raster(pass, raster)
}
\arguments{
\item{raster}{}
}
\description{
Make distance to reef pass
}
