% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{make_dist_to_seagrass_raster}
\alias{make_dist_to_seagrass_raster}
\title{Make distance to seagrass raster}
\usage{
make_dist_to_seagrass_raster(cor, raster)
}
\arguments{
\item{raster}{}
}
\description{
Make distance to seagrass raster
}
