% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{make_dist_to_seagrass_raster_old}
\alias{make_dist_to_seagrass_raster_old}
\title{Make distance to seagrass raster OLD}
\usage{
make_dist_to_seagrass_raster_old(cor, raster)
}
\arguments{
\item{raster}{}
}
\description{
Make distance to seagrass raster OLD
}
