% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{make_mpa_type_raster_no_take_only}
\alias{make_mpa_type_raster_no_take_only}
\title{Make MPA type raster no-take only}
\usage{
make_mpa_type_raster_no_take_only(raster)
}
\arguments{
\item{raster}{}
}
\description{
Make MPA type raster no-take only
}
