% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{make_mpa_type_raster_no_take_only_poe_ile_verte}
\alias{make_mpa_type_raster_no_take_only_poe_ile_verte}
\title{Make MPA type raster no-take only poe / ile verte (in surveyed block)}
\usage{
make_mpa_type_raster_no_take_only_poe_ile_verte(raster)
}
\arguments{
\item{raster}{}
}
\description{
Make MPA type raster no-take only poe / ile verte (in surveyed block)
}
