% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{make_pop_density}
\alias{make_pop_density}
\title{Prepare population density raster (worldwide data, 2020 year and 5 x 5 km buffer)}
\usage{
make_pop_density(raster, lon1, lon2, lat1, lat2)
}
\arguments{
\item{raster}{}

\item{lon1}{}

\item{lon2}{}

\item{lat1}{}

\item{lat2}{}
}
\description{
Prepare population density raster (worldwide data, 2020 year and 5 x 5 km buffer)
}
