% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{make_population_raster}
\alias{make_population_raster}
\title{Make population raster}
\usage{
make_population_raster(raster)
}
\arguments{
\item{raster}{}
}
\description{
Make population raster
}
