% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{make_raster_coral_benthic_type}
\alias{make_raster_coral_benthic_type}
\title{Make coverage raster for each habitat type of Allen Atlas}
\usage{
make_raster_coral_benthic_type(cor, r_xy, hab_type)
}
\arguments{
\item{cor}{}

\item{r_xy}{}
}
\description{
Make coverage raster for each habitat type of Allen Atlas
}
