% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{make_raster_coral_geomorpho_type}
\alias{make_raster_coral_geomorpho_type}
\title{Make coverage raster for each geomorphology type}
\usage{
make_raster_coral_geomorpho_type(cor, r_xy, geom_type)
}
\arguments{
\item{cor}{}

\item{r_xy}{}

\item{geom_type}{}
}
\description{
Make coverage raster for each geomorphology type
}
