% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{make_raster_coral_millenium_type}
\alias{make_raster_coral_millenium_type}
\title{Make coverage raster for each geomorphology type of Millenium atlas}
\usage{
make_raster_coral_millenium_type(coral, r_xy, geom_type)
}
\arguments{
\item{r_xy}{}

\item{geom_type}{}

\item{cor}{}
}
\description{
Make coverage raster for each geomorphology type of Millenium atlas
}
