% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{make_raster_seagrass}
\alias{make_raster_seagrass}
\title{Make coverage raster for seagrass}
\usage{
make_raster_seagrass(seagrass, r_xy)
}
\arguments{
\item{seagrass}{}

\item{r_xy}{}
}
\description{
Make coverage raster for seagrass
}
