% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{map_abundance_predictions_region}
\alias{map_abundance_predictions_region}
\title{Map abundance predictions, st errors and coef variations in region}
\usage{
map_abundance_predictions_region(maplatlonproj, pred, pred_se)
}
\arguments{
\item{pred}{}

\item{pred_se}{}
}
\description{
Map abundance predictions, st errors and coef variations in region
}
