% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{map_abundance_predictions_region_interp_mpas}
\alias{map_abundance_predictions_region_interp_mpas}
\title{Map abundance predictions, st errors and coef variations in region in univariate interpolation situations with mpas}
\usage{
map_abundance_predictions_region_interp_mpas(maplatlonproj, pred, pred_se, pa)
}
\arguments{
\item{pred}{}

\item{pred_se}{}
}
\description{
Map abundance predictions, st errors and coef variations in region in univariate interpolation situations with mpas
}
