% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{map_abundance_predictions_region_interp_new}
\alias{map_abundance_predictions_region_interp_new}
\title{Map abundance predictions new}
\usage{
map_abundance_predictions_region_interp_new(maplatlonproj, pred)
}
\arguments{
\item{maplatlonproj}{}

\item{pred}{}
}
\description{
Map abundance predictions new
}
