% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{map_abundance_predictions_surveyed}
\alias{map_abundance_predictions_surveyed}
\title{Map abundance predictions, st errors and coef variations in surveyed area}
\usage{
map_abundance_predictions_surveyed(maplatlonproj, pred, pred_se)
}
\arguments{
\item{maplatlonproj}{}

\item{pred}{}
}
\description{
Map abundance predictions, st errors and coef variations in surveyed area
}
