% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basic_maps.R
\name{map_allen_coral_osm}
\alias{map_allen_coral_osm}
\title{Make map of Allen coral geomorphology polygons overlaid on osm}
\usage{
map_allen_coral_osm(
  maplatlon,
  pa,
  cor,
  lon1,
  lon2,
  lat1,
  lat2,
  dist,
  offset_lon,
  offset_lat
)
}
\arguments{
\item{dist}{}
}
\description{
Make map of Allen coral geomorphology polygons overlaid on osm
}
