% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_density_maps.R
\name{map_dens_per_grid_species_log_with_zero}
\alias{map_dens_per_grid_species_log_with_zero}
\title{Map density per grid cell for given species (nb of observations / (length of tracks in meters * footprint_width in meters))}
\usage{
map_dens_per_grid_species_log_with_zero(
  maplatlonproj,
  polyobs,
  polytracks,
  footprintwidth,
  species,
  size
)
}
\arguments{
\item{size}{}
}
\description{
Map density per grid cell for given species (nb of observations / (length of tracks in meters * footprint_width in meters))
}
