% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{map_dist_to_all_reef}
\alias{map_dist_to_all_reef}
\title{Map distance to all reef raster}
\usage{
map_dist_to_all_reef(raster)
}
\arguments{
\item{raster}{}
}
\description{
Map distance to all reef raster
}
