% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{map_dist_to_seagrass}
\alias{map_dist_to_seagrass}
\title{Map distance to seagrass raster}
\usage{
map_dist_to_seagrass(raster)
}
\arguments{
\item{raster}{}
}
\description{
Map distance to seagrass raster
}
