% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basic_maps.R
\name{map_indiv_dugong_per_image_stage_telemetry}
\alias{map_indiv_dugong_per_image_stage_telemetry}
\title{Make map of telemetry with individual dugong observations per image differenciating stage (adult vs juvenile)}
\usage{
map_indiv_dugong_per_image_stage_telemetry(maplatlon, telem, telem_obs)
}
\arguments{
\item{telem_obs}{}
}
\description{
Make map of telemetry with individual dugong observations per image differenciating stage (adult vs juvenile)
}
