% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basic_maps.R
\name{map_indiv_dugong_per_image_telemetry_no_take_mpas}
\alias{map_indiv_dugong_per_image_telemetry_no_take_mpas}
\title{Make map of telemetry with individual dugong observations per image with no take mpas}
\usage{
map_indiv_dugong_per_image_telemetry_no_take_mpas(
  maplatlon,
  telem,
  telem_obs,
  pa
)
}
\arguments{
\item{telem_obs}{}
}
\description{
Make map of telemetry with individual dugong observations per image with no take mpas
}
