% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basic_maps.R
\name{map_indiv_species_telemetry}
\alias{map_indiv_species_telemetry}
\title{Make map of telemetry with individual species observations}
\usage{
map_indiv_species_telemetry(maplatlon, telem, telem_obs)
}
\arguments{
\item{telem_obs}{}
}
\description{
Make map of telemetry with individual species observations
}
