% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basic_maps.R
\name{map_indiv_species_telemetry_separate}
\alias{map_indiv_species_telemetry_separate}
\title{Make map of telemetry with individual species observations with separate map per species}
\usage{
map_indiv_species_telemetry_separate(maplatlon, telem, telem_obs)
}
\arguments{
\item{telem_obs}{}
}
\description{
Make map of telemetry with individual species observations with separate map per species
}
