% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{map_mpa_type_no_take_only}
\alias{map_mpa_type_no_take_only}
\title{Map mpa type no take only}
\usage{
map_mpa_type_no_take_only(raster)
}
\arguments{
\item{raster}{}
}
\description{
Map mpa type no take only
}
