% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basic_maps.R
\name{map_new_caledonia_big_mpas}
\alias{map_new_caledonia_big_mpas}
\title{Make New caledonia map with Poé location}
\usage{
map_new_caledonia_big_mpas(mpas, transects)
}
\description{
Make New caledonia map with Poé location
}
