% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basic_maps.R
\name{map_new_caledonia_surveyed_block}
\alias{map_new_caledonia_surveyed_block}
\title{make basic map of New caledonia weith surveyed block}
\usage{
map_new_caledonia_surveyed_block(nc, block)
}
\arguments{
\item{nc}{}

\item{block}{}
}
\description{
make basic map of New caledonia weith surveyed block
}
