% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_density_maps.R
\name{map_obs_per_grid_per_date_species}
\alias{map_obs_per_grid_per_date_species}
\title{Map number of observations per grid cell for given species}
\usage{
map_obs_per_grid_per_date_species(maplatlonproj, polygon, species, size)
}
\arguments{
\item{species}{}

\item{size}{}
}
\description{
Map number of observations per grid cell for given species
}
