% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_density_maps.R
\name{map_obs_per_grid_per_date_species_with_zero}
\alias{map_obs_per_grid_per_date_species_with_zero}
\title{Map number of observations per grid cell per date for given species with zeros}
\usage{
map_obs_per_grid_per_date_species_with_zero(
  maplatlonproj,
  polygon,
  polytracks,
  species,
  size
)
}
\arguments{
\item{species}{}

\item{size}{}
}
\description{
Map number of observations per grid cell per date for given species with zeros
}
