% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve_prioritisation.R
\name{map_reserves_minset_no_penalty_target_osm_abundance}
\alias{map_reserves_minset_no_penalty_target_osm_abundance}
\title{Map reserves with osm map and overlay of abundance predictions}
\usage{
map_reserves_minset_no_penalty_target_osm_abundance(
  maplatlonproj,
  pred,
  r,
  mpa,
  t
)
}
\arguments{
\item{pred}{}

\item{r}{}

\item{mpa}{}

\item{t}{}

\item{p}{}
}
\description{
Map reserves with osm map and overlay of abundance predictions
}
