% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve_prioritisation.R
\name{map_reserves_minset_no_penalty_target_osm_abundance_new_multi}
\alias{map_reserves_minset_no_penalty_target_osm_abundance_new_multi}
\title{Map reserves with osm map and overlay of abundance predictions multimap}
\usage{
map_reserves_minset_no_penalty_target_osm_abundance_new_multi(
  maplatlonproj,
  r30,
  r50,
  r80,
  t
)
}
\arguments{
\item{maplatlonproj}{}

\item{r30}{}

\item{r50}{}

\item{r80}{}

\item{t}{}
}
\description{
Map reserves with osm map and overlay of abundance predictions multimap
}
