% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve_prioritisation.R
\name{map_reserves_minset_with_penalty_target_lockin_osm}
\alias{map_reserves_minset_with_penalty_target_lockin_osm}
\title{Map reserves with osm map}
\usage{
map_reserves_minset_with_penalty_target_lockin_osm(maplatlonproj, r, p, t)
}
\arguments{
\item{pred}{}

\item{pred_se}{}
}
\description{
Map reserves with osm map
}
