% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_density_maps.R
\name{map_surveyed_area_per_grid_species_with_zeros}
\alias{map_surveyed_area_per_grid_species_with_zeros}
\title{Map surveyed area per grid cell for given species with zeros}
\usage{
map_surveyed_area_per_grid_species_with_zeros(
  maplatlonproj,
  raster_abund,
  species,
  size
)
}
\arguments{
\item{raster_abund}{}

\item{size}{}

\item{telem}{}
}
\description{
Map surveyed area per grid cell for given species with zeros
}
