% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basic_maps.R
\name{map_telemetry_date}
\alias{map_telemetry_date}
\title{Make map of telemetry with one color per date}
\usage{
map_telemetry_date(maplatlon, telem)
}
\arguments{
\item{maplatlon}{}
}
\description{
Make map of telemetry with one color per date
}
