% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basic_maps.R
\name{map_telemetry_date_separate}
\alias{map_telemetry_date_separate}
\title{Make separate map of telemetry per date}
\usage{
map_telemetry_date_separate(maplatlon, telem)
}
\arguments{
\item{maplatlon}{}
}
\description{
Make separate map of telemetry per date
}
