% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basic_maps.R
\name{map_telemetry_transects}
\alias{map_telemetry_transects}
\title{Make map of telemetry and transect lines}
\usage{
map_telemetry_transects(maplatlon, telem, transects)
}
\arguments{
\item{transects}{}
}
\description{
Make map of telemetry and transect lines
}
