% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_density_maps.R
\name{map_tracklen_per_grid_per_date}
\alias{map_tracklen_per_grid_per_date}
\title{Map track length per grid cell per date}
\usage{
map_tracklen_per_grid_per_date(maplatlonproj, polytracks, size)
}
\arguments{
\item{size}{}
}
\description{
Map track length per grid cell per date
}
