% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basic_maps.R
\name{map_transects_mpas}
\alias{map_transects_mpas}
\title{Make map of selected transect lines with mpas}
\usage{
map_transects_mpas(maplatlon, transects, pa)
}
\arguments{
\item{transects}{}
}
\description{
Make map of selected transect lines with mpas
}
