% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basic_maps.R
\name{map_transects_notake_mpas}
\alias{map_transects_notake_mpas}
\title{Make map of selected transect lines with no take mpas}
\usage{
map_transects_notake_mpas(maplatlon, transects, pa)
}
\arguments{
\item{transects}{}
}
\description{
Make map of selected transect lines with no take mpas
}
