% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{mask_with_land2}
\alias{mask_with_land2}
\title{Mask based on dist to mainland raster}
\usage{
mask_with_land2(r, mask)
}
\arguments{
\item{r}{}

\item{mask}{}
}
\description{
Mask based on dist to mainland raster
}
