% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basic_maps.R
\name{osm_map}
\alias{osm_map}
\title{Make open street map for study area (satellite view)}
\usage{
osm_map(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lon2}{}
}
\description{
Make open street map for study area (satellite view)
}
