% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{read_average_resample_turbidity}
\alias{read_average_resample_turbidity}
\title{Read New Caledonia Allen coral quaterly turbidity, produce average raster and resample}
\usage{
read_average_resample_turbidity(lon1, lon2, lat2, lat1, rast)
}
\arguments{
\item{lat1}{}
}
\description{
Read New Caledonia Allen coral quaterly turbidity, produce average raster and resample
}
