% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{read_convert_coral_benthic}
\alias{read_convert_coral_benthic}
\title{Read and convert New Caledonia Allen coral shapefile benthic}
\usage{
read_convert_coral_benthic(lon1, lon2, lat2, lat1)
}
\arguments{
\item{lat1}{}
}
\description{
Read and convert New Caledonia Allen coral shapefile benthic
}
