% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{read_convert_coral_geomorpho}
\alias{read_convert_coral_geomorpho}
\title{Read and convert New Caledonia Allen coral shapefile geomorphology}
\usage{
read_convert_coral_geomorpho(lon1, lon2, lat2, lat1)
}
\arguments{
\item{lat1}{}
}
\description{
Read and convert New Caledonia Allen coral shapefile geomorphology
}
