% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{read_convert_seagrass_nc}
\alias{read_convert_seagrass_nc}
\title{Read seagrass New Caledonia from Andrefouet et al 2021 Marine Pollution Bulletin}
\usage{
read_convert_seagrass_nc()
}
\description{
Read seagrass New Caledonia from Andrefouet et al 2021 Marine Pollution Bulletin
}
