% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{read_original_coralnc_millenium}
\alias{read_original_coralnc_millenium}
\title{Read New Caledonia coral shapefile (milenium coral project)}
\usage{
read_original_coralnc_millenium()
}
\description{
Read New Caledonia coral shapefile (milenium coral project)
}
