% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{read_project_surveyed_block}
\alias{read_project_surveyed_block}
\title{read and project surveyed block (created in qgis)}
\usage{
read_project_surveyed_block()
}
\description{
read and project surveyed block (created in qgis)
}
