% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{read_project_surveyed_strata}
\alias{read_project_surveyed_strata}
\title{read and project surveyed strata (created in qgis)}
\usage{
read_project_surveyed_strata(name)
}
\arguments{
\item{name}{}
}
\description{
read and project surveyed strata (created in qgis)
}
