% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{read_resample_depth}
\alias{read_resample_depth}
\title{read depth from Jean roger (IRD) and resample to study area}
\usage{
read_resample_depth(rastlatlon, rastxy)
}
\arguments{
\item{rast}{}
}
\description{
read depth from Jean roger (IRD) and resample to study area
}
