% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_clean.R
\name{select_obs_telemetry_wcoast}
\alias{select_obs_telemetry_wcoast}
\title{select observations for west coast}
\usage{
select_obs_telemetry_wcoast(telem_obs, ls_vids)
}
\arguments{
\item{ls_vids}{}
}
\description{
select observations for west coast
}
