% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_clean.R
\name{select_on_effort_wcoast}
\alias{select_on_effort_wcoast}
\title{Select on effort data (telemetry or observations) west coast}
\usage{
select_on_effort_wcoast(telem, off_transit, off_loop)
}
\arguments{
\item{off}{}
}
\description{
Select on effort data (telemetry or observations) west coast
}
