% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_clean.R
\name{select_telemetry_wcoast}
\alias{select_telemetry_wcoast}
\title{select telemetry for west coast}
\usage{
select_telemetry_wcoast(telem, ls_vids)
}
\arguments{
\item{ls_vids}{}
}
\description{
select telemetry for west coast
}
