% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{fit_full_abundance_model_k5}
\alias{fit_full_abundance_model_k5}
\title{fit full abudance GAM with area surveyed as offset and
automated term selection thanks to select = T (https://stats.stackexchange.com/questions/405129/model-selection-for-gam-in-r)
with k limited to 5}
\usage{
fit_full_abundance_model_k5(distrib, data_model)
}
\arguments{
\item{distrib}{}
}
\description{
fit full abudance GAM with area surveyed as offset and
automated term selection thanks to select = T (https://stats.stackexchange.com/questions/405129/model-selection-for-gam-in-r)
with k limited to 5
}
